Add BSD support and remove duplicated code.
