+++ report/os_xdg.go
@@ -1,3 +1,4 @@
+// +build linux openbsd freebsd netbsd dragonfly
 package report
 
 import (
@@ -5,7 +6,6 @@ import (
 	"bytes"
 	"fmt"
 	"io/ioutil"
-	"os/exec"
 	"path/filepath"
 	"strings"
 
@@ -74,28 +74,4 @@ func (i *OS) parseOsReleaseCmdOutput(data []byte) (goinfo.Info, error) {
 		info[key] = strings.Trim(tokens[1], `"`)
 	}
 	return info, scanner.Err()
-}
-
-func (i *OS) architecture() (string, error) {
-	cmd := exec.Command("uname", "-m")
-	out, err := cmd.Output()
-	if err != nil {
-		return "", fmt.Errorf("could not detect architecture using uname command: %w", err)
-	}
-
-	info := strings.Trim(string(out), "\n")
-
-	return info, nil
-}
-
-func (i *OS) kernel() (string, error) {
-	cmd := exec.Command("uname", "-rsv")
-	out, err := cmd.Output()
-	if err != nil {
-		return "", fmt.Errorf("could not detect the kernel using uname command: %w", err)
-	}
-
-	info := strings.Trim(string(out), "\n")
-
-	return info, nil
 }
